Addressed PR comments, moved UniformHash to fuel
