+++ nvflare/edge/web/routing_proxy.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import hashlib
 import json
 import logging
 import os
@@ -24,46 +23,10 @@ from flask import Flask, Response, jsonify, request
 
 from nvflare.edge.web.models.api_error import ApiError
 from nvflare.edge.web.web_server import FilteredJSONProvider
-
-# This is just a random large prime number used as number of buckets
-PRIME = 100003
+from nvflare.fuel.utils.hash_utils import UniformHash
 
 log = logging.getLogger(__name__)
 app = Flask(__name__)
-
-
-class UniformHash:
-    """A hash algorithm with uniform distribution. It achieves this with following steps,
-    1. Get a hash value using SHA256
-    2. Map the hash value to a virtual hash table with a large prime number
-    3. Map the virtual bucket to real bucket by using an allocation table
-
-    """
-
-    def __init__(self, num_buckets: int):
-        self.num_buckets = num_buckets
-        self.num = PRIME // num_buckets
-        self.remainder = PRIME % num_buckets
-
-    def get_num_buckets(self) -> int:
-        return self.num_buckets
-
-    def hash(self, key: str) -> int:
-        # The hash() function changes value every run so SHA256 is used
-        sha_bytes = hashlib.sha256(key.encode()).digest()
-        sha = int.from_bytes(sha_bytes[:8], "big")
-        virtual_hash = sha % PRIME
-
-        start = 0
-        for i in range(self.num_buckets):
-            # Allocation is virtual hash assigned to each bucket, first few buckets get one more if r is not 0
-            allocation = (self.num + 1) if i < self.remainder else self.num
-            end = start + allocation
-            if start <= virtual_hash < end:
-                return i
-            start = end
-
-        raise RuntimeError("Logic error")
 
 
 class LcpMapper:
